requirements.txt
snowflake-snowpark-python[pandas]
streamlit==1.20.0